Move protocol definition to separate module
